garmin integration complete + persistance

+++ fitness-aggregator-api/src/main/java/com/jasonghent98/fitness_aggregator_api/dto/garmin/webhook/GarminHrvSummaryPayload.java
@@ -37,10 +37,10 @@ public class GarminHrvSummaryPayload {
         private Integer startTimeOffsetInSeconds;
 
         @JsonProperty("lastNightAvg")
-        private Double lastNightAvg;
+        private Integer lastNightAvg;
 
         @JsonProperty("lastNight5MinHigh")
-        private Double lastNight5MinHigh;
+        private Integer lastNight5MinHigh;
 
         @JsonProperty("hrvValues")
         private Map<String, Double> hrvValues;  // keys are offset in seconds, values are HRV values
